# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Laksamana API v1 — account, jadwal, marketing",
    "version": "1.0.0",
    "description": "Hand-written contract for the `/api/v1` surface of the account, jadwal and marketing modules. Sanctum Bearer auth; the shared `{data, meta}` / `{error:{code,message,details}}` envelopes (App\\Support\\Api\\ApiResponse) are the components below. The drift check `tools/openapi-check.php` keeps this list in step with `php artisan route:list --path=api/v1`."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum"
      }
    },
    "schemas": {
      "Success": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {},
          "meta": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "description": "The shared success envelope. `meta` carries versions (e.g. `version`, `versions`, `total`)."
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine code, e.g. `module_not_granted`, `not_found`, `forbidden`."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "UserWrite": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "pin": {
            "type": "string",
            "pattern": "^\\d{4,8}$"
          },
          "active": {
            "type": "boolean"
          },
          "keterangan": {
            "type": "string",
            "maxLength": 255
          },
          "noHp": {
            "type": "string",
            "maxLength": 32
          },
          "talentaId": {
            "type": "string",
            "maxLength": 64
          },
          "username": {
            "type": "string",
            "maxLength": 40
          },
          "branch": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "jobPosition": {
            "type": "string"
          },
          "jobLevel": {
            "type": "string"
          },
          "employmentStatus": {
            "type": "string"
          },
          "joinDate": {
            "type": "string"
          }
        }
      },
...
      "ModulesWrite": {
        "type": "object",
        "required": [
          "modules"
        ],
        "properties": {
          "modules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          }
        }
      },
      "Module": {
        "type": "object",
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "string",
            "maxLength": 64
          },
          "label": {
            "type": "string",
            "maxLength": 120
          },
          "active": {
            "type": "boolean"
          },
          "urutan": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/v1/account/users": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "List Office Users",
        "operationId": "getApiV1AccountUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (missing/invalid Bearer token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (module not granted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Create an Office User",
        "operationId": "postApiV1AccountUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (missing/invalid Bearer token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (module not granted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserWrite"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/account/modules": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "List modules",
        "operationId": "getApiV1AccountModules",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (missing/invalid Bearer token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (module not granted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Sync the module registry",
        "operationId": "postApiV1AccountModules",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (missing/invalid Bearer token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (module not granted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModulesWrite"
              }
            }
          }
        }
      }
    },
...
  }
}